# openapi.yaml
openapi: 3.1.0
info:
  title: 山海灵感便利店领域 API
  version: 2.0.0
  description: App 用户接口与管理端事务动作。字段统一 snake_case，金额统一十进制字符串。
servers:
  - url: http://localhost:3000/api/v1
...
paths:
...
  /auth/phone/login:
    post:
      tags: [Auth]
      summary: 手机验证码登录（当前仅非生产开发桩）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code]
              properties:
                phone: { type: string, pattern: '^1\d{10}$' }
                code: { type: string }
                device: { type: string }
                platform: { type: string, enum: [android, ios, pc] }
      responses: { "200": { $ref: "#/components/responses/Success" }, "401": { $ref: "#/components/responses/Error" } }
  /auth/oauth/login:
    post:
      tags: [Auth]
      summary: 微信或 Apple 登录（需配置验证适配器）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identity_type, identifier]
              properties:
                identity_type: { type: string, enum: [wechat, apple] }
                identifier: { type: string }
                credential: { type: string }
                phone: { type: string, description: 首次登录时必填 }
                device: { type: string }
                platform: { type: string, enum: [android, ios, pc] }
      responses: { "200": { $ref: "#/components/responses/Success" }, "503": { $ref: "#/components/responses/Error" } }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: 轮换用户会话
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [token], properties: { token: { type: string } } } } }
      responses: { "200": { $ref: "#/components/responses/Success" }, "401": { $ref: "#/components/responses/Error" } }
  /auth/logout:
    post:
      tags: [Auth]
      security: [{ UserBearer: [] }]
      summary: 退出当前会话
      responses: { "204": { description: 已退出 }, "401": { $ref: "#/components/responses/Error" } }
...
  /addresses:
    get:
      security: [{ UserBearer: [] }]
      summary: 本人地址列表
      responses: { "200": { $ref: "#/components/responses/Success" } }
    post:
      security: [{ UserBearer: [] }]
      summary: 新增地址；首个地址自动默认
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AddressInput" } } } }
      responses: { "201": { $ref: "#/components/responses/Success" } }
  /addresses/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    patch:
      security: [{ UserBearer: [] }]
      summary: 修改本人地址
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AddressInput" } } } }
      responses: { "200": { $ref: "#/components/responses/Success" } }
    delete:
      security: [{ UserBearer: [] }]
      summary: 删除本人地址并补选默认地址
      responses: { "204": { description: 已删除 } }
  /addresses/{id}/default:
    put:
      security: [{ UserBearer: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      summary: 事务设置默认地址
      responses: { "200": { $ref: "#/components/responses/Success" } }
...
  /admin/products/{id}/stock-adjustments:
    post:
      tags: [Admin]
      security: [{ AdminBearer: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      summary: 事务调整商品或 SKU 库存
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [change_qty, reason], properties: { change_qty: { type: integer }, sku_id: { $ref: "#/components/schemas/Id" }, reason: { type: string }, biz_id: { type: string }, operator_id: { $ref: "#/components/schemas/Id" } } } } } }
      responses: { "201": { $ref: "#/components/responses/Success" } }
  /admin/orders/{orderNo}/ship:
    post:
      tags: [Admin]
      security: [{ AdminBearer: [] }]
      parameters: [{ $ref: "#/components/parameters/OrderNo" }]
      summary: 发货并流转为待收货
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [carrier, tracking_no], properties: { carrier: { type: string }, tracking_no: { type: string }, operator_id: { $ref: "#/components/schemas/Id" } } } } } }
      responses: { "200": { $ref: "#/components/responses/Success" } }
  /admin/scan-products/{id}/approve:
    post:
      tags: [Admin]
      security: [{ AdminBearer: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      summary: 审核通过并创建正式商品
      requestBody: { content: { application/json: { schema: { type: object, properties: { stock: { type: integer, minimum: 0 }, reviewer_id: { $ref: "#/components/schemas/Id" } } } } } }
      responses: { "200": { $ref: "#/components/responses/Success" } }
  /admin/scan-products/{id}/reject:
    post:
      tags: [Admin]
      security: [{ AdminBearer: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      summary: 拒绝扫码商品
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [reason], properties: { reason: { type: string }, reviewer_id: { $ref: "#/components/schemas/Id" } } } } } }
      responses: { "200": { $ref: "#/components/responses/Success" } }
...
components:
  securitySchemes:
    UserBearer: { type: http, scheme: bearer, description: 登录返回的 App 会话 Token }
    AdminBearer: { type: http, scheme: bearer, description: 环境变量 API_TOKEN }
  parameters:
    Id: { name: id, in: path, required: true, schema: { $ref: "#/components/schemas/Id" } }
    OrderNo: { name: orderNo, in: path, required: true, schema: { type: string } }
...
  schemas:
    Id: { oneOf: [{ type: string, pattern: '^\d+$' }, { type: integer, minimum: 1 }] }
...
    AddressInput:
      type: object
      properties:
        consignee: { type: string, maxLength: 32 }
        phone: { type: string, pattern: '^1\d{10}$' }
        province: { type: string }
        city: { type: string }
        district: { type: string }
        detail: { type: string, maxLength: 255 }
        tag: { type: [string, "null"] }
        is_default: { type: boolean }
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties: { code: { type: string }, message: { type: string }, details: {} }
        requestId: { type: string }
  responses:
    Success:
      description: 成功
      content: { application/json: { schema: { type: object, properties: { data: {} } } } }
...
    Error:
      description: 请求失败
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }